+++ src/components/layout/root-layout.tsx
@@ -1,6 +1,6 @@
-import Header from '@components/layout/Header/Header';
-import Main from '@components/layout/Main';
-import Footer from '@components/layout/Footer/Footer';
+import Header from '@components/layout/Header/header';
+import Main from '@components/layout/main';
+import Footer from '@components/layout/Footer/footer';
 import { Toaster } from '@components/ui/toaster';
 
 export default function RootLayout({
